pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "project"
authors = [
    {name = "IonWorks Technologies Inc.", email = "info@ion-works.com"},
]
description = "Code provided by IonWorks Technologies Inc."
readme = "README.md"
dynamic = ["version"]
requires-python = ">=3.8,<3.12"
dependencies = [
    "pybamm",
    "python-dotenv",
]

[project.entry-points.pybamm_parameter_sets]
MyCell = "projectname.parameters.mycell:get_parameter_values"
